requirements.txt
Django
djangorestframework
django-reversion > 3.0.7
dj-database-url
